brincus-education-analysis/requirements.txt
# API and Data Processing
requests==2.31.0
pandas==2.1.0
numpy==1.24.3

# PDF Generation
reportlab==4.0.4
WeasyPrint==59.0
Jinja2==3.1.2

# Progress and CLI
tqdm==4.65.0
click==8.1.7

# Development
python-dotenv==1.0.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0

# Logging and Monitoring
structlog==23.1.0